fix(review): retain checked base revision
